Improved widget passthrough for experience cards
