+++ lib/data.dart
@@ -1,19 +1,4 @@
-import 'package:flutter/painting.dart';
-
-class VideoContent
-{
-  VideoContent(String newTitle, String newDescription, String newThumbnail, String newUrl)
-  {
-    this.title = newTitle;
-    this.description = newDescription;
-    this.thumbnail = new AssetImage(newThumbnail);
-    this.videoUrl = newUrl;
-  }
-  String title;
-  String description;
-  AssetImage thumbnail;
-  String videoUrl;
-}
+import 'video_content.dart';
 
 final List<VideoContent> data = <VideoContent>[
   VideoContent("Mule",
